Move `ByteCount` into `Methods` lint pass
